requirements.txt
keyboard<=0.13.5
PyYAML<=5.3.1